Added unlink image from edit and delete employee API.

+++ backend/app/Http/Controllers/EmployeeController.php
@@ -142,7 +142,9 @@ class EmployeeController extends Controller
                 $name = time().'.' . explode('/', explode(':', substr($image, 0, strpos($image, ';')))[1])[1];
                 \Storage::disk('local')->putFileAs(config('config.user.profile_image_path'), $image, $name);
 
-                unlink(public_path(config('config.user.profile_image_path'). $user->profile));
+                if ($user->profile) {
+                    unlink(public_path(config('config.user.profile_image_path'). $user->profile));
+                }
                 $user->profile = $name;
             }
             $user->description = $request->description;
@@ -172,7 +174,9 @@ class EmployeeController extends Controller
             return $this->error("ADMIN_DELETE");
         }
         try {
-            unlink(public_path(config('config.user.profile_image_path'). $user->profile));
+            if ($user->profile) {
+                unlink(public_path(config('config.user.profile_image_path'). $user->profile));
+            }
             if ($user->delete()) {
                 return $this->success([], "EMPLOYEE_DELETED");
             } else {
